requirements.txt
fastapi
uvicorn
pydantic
python-multipart
psycopg2-binary
opencv-python
ultralytics
python-dotenv
PyJWT
